+++ src/board/transition.ts
@@ -10,7 +10,8 @@ import { LayoutEngine } from "../internal/layout-engine/engine";
 import { Coordinates } from "../internal/utils/coordinates";
 import { getDefaultColumnSpan, getDefaultRowSpan, getMinColumnSpan, getMinRowSpan } from "../internal/utils/layout";
 import { Position } from "../internal/utils/position";
-import { BoardProps, RemoveTransition, Transition, TransitionAnnouncement } from "./interfaces";
+import { BoardProps } from "./interfaces";
+import { RemoveTransition, Transition, TransitionAnnouncement } from "./internal-interfaces";
 import { createOperationAnnouncement } from "./utils/announcements";
 import { getHoveredRect } from "./utils/get-hovered-rect";
 import { getInsertionDirection, getLayoutPlaceholders, getLayoutRows, getLayoutShift } from "./utils/layout";
